+++ app/build.gradle
@@ -35,6 +35,7 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.5.0'
     implementation 'com.github.bumptech.glide:glide:4.10.0'
 
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
@@ -42,3 +43,4 @@ dependencies {
     implementation 'androidx.cardview:cardview:1.0.0'
 
 }
+apply plugin: 'kotlin-android-extensions'
