+++ app/src/main/java/com/example/wenesdaytaskkotlin/Adapters/ArtistAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.wenesdaytaskkotlin
+package com.example.wenesdaytaskkotlin.Adapters
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -8,26 +8,32 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.wenesdaytaskkotlin.R
+import com.example.wenesdaytaskkotlin.Models.Result
+import kotlinx.android.synthetic.main.custom_grid_layout.view.*
 
-class ArtistAdapter(
-
-    context: Context?,
-    resultList: List<Result>
-
-) : RecyclerView.Adapter<ArtistAdapter.ViewHolder>() {
+class ArtistAdapter(context: Context?, resultList: List<Result>) : RecyclerView.Adapter<ArtistAdapter.ViewHolder>() {
 
     var layoutInflater: LayoutInflater
     var resultList: List<Result>
 
+    init {
+        layoutInflater = LayoutInflater.from(context)
+        this.resultList = resultList
+    }
+
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+
         val view = layoutInflater.inflate(R.layout.custom_grid_layout, parent, false)
 
         return ViewHolder(view)
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+
         Glide.with(layoutInflater.context).load(resultList[position].artworkUrl30)
             .into(holder.artistImg)
+
         holder.name.text = resultList[position].artistName
     }
 
@@ -35,19 +41,12 @@ class ArtistAdapter(
         return resultList.size
     }
 
-    inner class ViewHolder(itemView: View) :
-        RecyclerView.ViewHolder(itemView) {
-        var artistImg: ImageView
-        var name: TextView
+    inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
-        init {
-            artistImg = itemView.findViewById(R.id.img)
-            name = itemView.findViewById(R.id.name)
-        }
+        var artistImg: ImageView = itemView.image_view_artists
+        var name: TextView = itemView.text_view_name
+
     }
 
-    init {
-        layoutInflater = LayoutInflater.from(context)
-        this.resultList = resultList
-    }
+
 }
