+++ app/src/main/java/com/example/wenesdaytaskkotlin/Models/Result.kt
@@ -1,4 +1,4 @@
-package com.example.wenesdaytaskkotlin
+package com.example.wenesdaytaskkotlin.Models
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
