+++ app/src/main/res/layout/custom_grid_layout.xml
@@ -22,7 +22,7 @@
             android:weightSum="4">
 
             <ImageView
-                android:id="@+id/img"
+                android:id="@+id/image_view_artists"
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
                 android:layout_centerHorizontal="true"
@@ -32,10 +32,10 @@
                 app:srcCompat="@drawable/ic_launcher_background" />
 
             <TextView
-                android:id="@+id/name"
+                android:id="@+id/text_view_name"
                 android:layout_width="wrap_content"
                 android:layout_height="0dp"
-                android:layout_below="@id/img"
+                android:layout_below="@id/image_view_artists"
                 android:layout_centerHorizontal="false"
                 android:layout_gravity="center"
                 android:layout_marginStart="5dp"
